refactor: Update modules and imports for code organization and consistency
