--- dorm/src/expression/mod.rs
@@ -1,6 +1,0 @@
-pub mod expression;
-pub mod expression_arc;
-
-pub use crate::expression::expression::Expression;
-pub use crate::expression::expression_arc::ExpressionArc;
-pub use crate::expression::expression_arc::WrapArc;
