+++ dorm/src/prelude.rs
@@ -2,13 +2,13 @@ pub use crate::dataset::ReadableDataSet;
 pub use crate::datasource::postgres::*;
 pub use crate::expr;
 pub use crate::expr_arc;
-pub use crate::expression::expression_arc::WrapArc;
-pub use crate::expression::{Expression, ExpressionArc};
 pub use crate::field::Field;
 pub use crate::operations::Operations;
+pub use crate::sql::chunk::SqlChunk;
+pub use crate::sql::expression::{Expression, ExpressionArc};
+pub use crate::sql::WrapArc;
 pub use crate::table::TableDelegate;
 pub use crate::traits::any::AnyTable;
 pub use crate::traits::any::RelatedTable;
 pub use crate::traits::entity::{EmptyEntity, Entity};
-pub use crate::traits::sql_chunk::SqlChunk;
 pub use crate::{query::JoinQuery, query::Query, table::Table};
