+++ dorm/src/sql/expression/expression.rs
@@ -1,9 +1,17 @@
 use serde_json::Value;
 
-use crate::{
-    operations::Operations,
-    traits::{column::Column, sql_chunk::SqlChunk},
-};
+use crate::{operations::Operations, sql::chunk::SqlChunk, traits::column::Column};
+
+/// Constructs [`Expression`] from a format scring and several parameters by passing those
+/// into [`json!`]
+///
+/// ```
+/// let my_sum = expr!("{} + {}", 2, 3);
+/// ```
+///
+/// The parameter to the expr! can be anything that you can also pass into [`json!`] macro
+///
+/// [`json!`]: serde_json::json!
 
 #[macro_export]
 macro_rules! expr {
@@ -17,12 +25,20 @@ macro_rules! expr {
     }}
 }
 
+/// Expression is a basic piece of SQL query template that contains a format string
+/// and several parameters. Easiest way to create Expression is with [`expr!`] macro
+///
+
 #[derive(Debug, Clone)]
 pub struct Expression {
     expression: String,
     parameters: Vec<Value>,
 }
 
+/// Expression can be used anywhere, where SqlChunk is accepted. For example:
+/// ```
+/// let expression = expr_arc!("{} + ({})", 2, expr!("3 * 4"));
+/// ```
 impl SqlChunk for Expression {
     fn render_chunk(&self) -> Expression {
         self.clone()
@@ -48,10 +64,17 @@ impl Expression {
         }
     }
 
+    /// Return "SQL" template part of the expression
     pub fn sql(&self) -> &String {
         &self.expression
     }
 
+    /// Converts template by replacing corresponding {} placeholders with $1, $2 etc,
+    /// which is more suitable for some SQL crates
+    ///
+    /// ```
+    /// let final = expr!("{} + {}", 2, 3);  // "$1 + $2"
+    /// ```
     pub fn sql_final(&self) -> String {
         let mut sql_final = self.expression.clone();
 
@@ -68,6 +91,15 @@ impl Expression {
         &self.parameters
     }
 
+    /// Given a Vec<Expression> and a delimeter, will construct a new expression,
+    /// by combining all nested templates together:
+    /// ```
+    /// let e1 = expr!("hello {}", "world");
+    /// let e2 = expr!("foo {}", "bar");
+    /// let e = Expression::from_vec(vec![e1, e2], " <=> ");
+    ///
+    /// writeln(e.sql()); // hello {} <=> foo {}
+    /// ```
     pub fn from_vec(vec: Vec<Expression>, delimiter: &str) -> Self {
         let expression = vec
             .iter()
@@ -87,10 +119,13 @@ impl Expression {
         }
     }
 
+    /// Return SQL template and parameter vec as a tuple
     pub fn split(self) -> (String, Vec<Value>) {
         (self.expression, self.parameters)
     }
 
+    /// Places values into the template and returns a String.
+    /// Useful for debugging, but not for SQL execution.
     pub fn preview(&self) -> String {
         let mut preview = self.expression.clone();
         for param in &self.parameters {
@@ -118,15 +153,21 @@ impl Operations for Expression {}
 
 #[cfg(test)]
 mod tests {
+    use super::*;
+    use crate::sql::chunk::SqlChunk;
+    use serde_json::json;
+
     #[test]
     fn test_as_type() {
-        use super::*;
-        use crate::traits::sql_chunk::SqlChunk;
-        use serde_json::json;
-
         let expression = Expression::as_type(json!(1), "int");
         let (sql, params) = expression.render_chunk().split();
         assert_eq!(sql, "{}::int");
         assert_eq!(params, vec![Value::Number(1.into())]);
+    }
+
+    #[test]
+    fn test_expr_macro() {
+        let expr = expr!("{} + {}", 2, 2);
+        assert_eq!(expr.preview(), "2 + 2");
     }
 }
