+++ dorm/src/traits/mod.rs
@@ -4,5 +4,3 @@ pub mod dataset;
 pub mod datasource;
 pub mod entity;
 // pub mod postgres;
-
-pub mod sql_chunk;
